+++ lib/widgets/collection/app_bar.dart
@@ -393,12 +393,7 @@ class _CollectionAppBarState extends State<CollectionAppBar> with RouteAware, Si
     required bool Function(EntrySetAction action) isVisible,
     required bool Function(EntrySetAction action) canApply,
   }) {
-    final isSelecting = selection.isSelecting;
-
-    return [
-      ...EntrySetActions.general,
-      ...isSelecting ? EntrySetActions.pageSelection : EntrySetActions.pageBrowsing,
-    ].nonNulls.where(isVisible).map((action) {
+    Widget toCaptionedButton(action) {
       final enabled = canApply(action);
       return CaptionedButton(
         iconButtonBuilder: (context, focusNode) => _buildButtonIcon(
@@ -411,7 +406,33 @@ class _CollectionAppBarState extends State<CollectionAppBar> with RouteAware, Si
         captionText: _buildButtonCaption(context, action, enabled: enabled),
         onPressed: enabled ? () => _onActionSelected(action) : null,
       );
-    }).toList();
+    }
+
+    final fabAction = EntrySetActions.fab.firstWhereOrNull(isVisible);
+
+    final isSelecting = selection.isSelecting;
+    final regularActions = [
+      ...EntrySetActions.general,
+      ...isSelecting ? EntrySetActions.pageSelection : EntrySetActions.pageBrowsing,
+    ].nonNulls.where(isVisible).map(toCaptionedButton).toList();
+
+    return [
+      if (fabAction != null) ...[
+        toCaptionedButton(fabAction),
+        const Align(
+          alignment: Alignment.topCenter,
+          child: SizedBox(
+            width: 16,
+            height: kMinInteractiveDimension,
+            child: Divider(
+              thickness: 4,
+              radius: BorderRadius.all(Radius.circular(123)),
+            ),
+          ),
+        ),
+      ],
+      ...regularActions,
+    ];
   }
 
   static double _iconButtonWidth(BuildContext context) {
@@ -770,6 +791,9 @@ class _CollectionAppBarState extends State<CollectionAppBar> with RouteAware, Si
       case .editRating:
       case .editTags:
       case .removeMetadata:
+      // fab
+      case .pickCollectionFilters:
+      case .pickMultipleMedia:
         _actionDelegate.onActionSelected(context, action);
     }
   }